Expose `folderToPackageId`
