+++ lib/plugins/dependency-tracker.js
@@ -7,6 +7,7 @@ import { DBPF, FileType, LotObjectType, ExemplarProperty } from 'sc4/core';
 import { hex } from 'sc4/utils';
 import PluginIndex from './plugin-index.js';
 import FileScanner from './file-scanner.js';
+import folderToPackageId from './folder-to-package-id.js';
 import * as Dep from './dependency-types.js';
 
 // Constants
@@ -695,18 +696,4 @@ class File extends String {
 		}
 		return opts.stylize(value, 'special');
 	}
-}
-
-// # folderToPackageId(folder)
-// Returns the corresponding sc4pac package id from a given folder. If this is 
-// not an sc4pac folder, we return nothing.
-function folderToPackageId(folder) {
-	let basename = path.basename(folder);
-	while (!basename.endsWith('.sc4pac')) {
-		folder = path.resolve(folder, '..');
-		basename = path.basename(folder);
-		if (!basename) return null;
-	}
-	let [group, name] = basename.split('.');
-	return `${group}:${name}`;
 }
