+++ lib/plugins/index.js
@@ -2,3 +2,4 @@
 export { default as PluginIndex } from './plugin-index.js';
 export { default as FileScanner } from './file-scanner.js';
 export { default as DependencyTracker } from './dependency-tracker.js';
+export { default as folderToPackageId } from './folder-to-package-id.js';
